requirements.txt
python-telegram-bot==20.4
pdf2docx==0.7.6
docx2pdf==0.1.8